Use GeoJSON; refactor

+++ src/main/scala/model/Event.scala
@@ -1,4 +1,4 @@
 package model
 
 case class Event(user_id: String, headline: String, cost: Float, duration: Int, description: String,
-    x: Long, y: Long)
+    x: Double, y: Double)
